backend/requirements.txt
fastapi
pydantic-settings
sqlalchemy
psycopg2-binary
uvicorn
alembic
dotenv
passlib[bcrypt]
PyJWT
python-multipart